Assets/Scripts/TransferScripts/requirements.txt
altgraph==0.17
certifi==2020.12.5
cycler==0.10.0
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
pefile==2019.4.18
Pillow==8.1.2
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pywin32-ctypes==0.2.0
readme-renderer==28.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
scipy==1.6.2
six==1.15.0
tqdm==4.60.0
twine==3.2.0
wincertstore==0.2
